requirements.txt
# Rex-Brain Requirements
# Python 3.10+ required

# ============================================
# ANDROID / KIVY FRAMEWORK
# ============================================
kivy==2.3.0
plyer==2.1.0                    # Access Android sensors (camera, mic, etc.)
pyjnius==1.6.1                  # Java bridge for Android

# ============================================
# ROBOT COMMUNICATION
# ============================================
aiortc==1.9.0                   # WebRTC for Python
aiohttp==3.9.1                  # Async HTTP client

# ============================================
# AUDIO PROCESSING
# ============================================
# Note: Using aiohttp for Deepgram WebSocket (better Android header support)
sounddevice==0.4.6              # Audio capture (desktop only)
numpy==1.26.2                   # Audio array processing

# ============================================
# VISION / IMAGE PROCESSING
# ============================================
opencv-python-headless==4.9.0.80  # Image processing (headless for Android)
mediapipe==0.10.9               # Face detection (lightweight, runs on device)
Pillow==10.2.0                  # Image manipulation

# ============================================
# LLM / AI APIS
# ============================================
# Note: Using direct HTTP calls (no SDK for Android compatibility)
httpx==0.27.0                   # HTTP client for Claude & OpenAI APIs

# ============================================
# MEMORY / DATABASE
# ============================================
sqlalchemy==2.0.23              # Database ORM
aiosqlite==0.19.0               # Async SQLite

# ============================================
# UTILITIES
# ============================================
pyyaml==6.0.1                   # Configuration files
python-dotenv==1.0.0            # Environment variables
pydantic==2.5.2                 # Data validation
pydantic-settings==2.1.0        # Settings management

# ============================================
# ASYNC / CONCURRENCY
# ============================================
asyncio-throttle==1.0.2         # Rate limiting for APIs
anyio==4.2.0                    # Async utilities (required by httpx)
sniffio==1.3.0                  # Async library detection

# ============================================
# DEVELOPMENT ONLY (not needed on Android)
# ============================================
# pytest==7.4.3
# pytest-asyncio==0.21.1

